requirements.txt
# freee API
requests>=2.31.0

# Gmail API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# LLM/OCR
anthropic>=0.18.0
pdf2image>=1.16.3
Pillow>=10.0.0

# Configuration
pyyaml>=6.0.1

# Utilities
python-dateutil>=2.8.2
typing-extensions>=4.8.0
